[core] Split `test_token_auth_integration.py` (#62267)
This test is periodically timing out on the Windows build. It is already
a medium test, so instead of bumping further, opting to split the test
in two to reduce the runtime in case of retries.

---------
