+++ python/ray/tests/BUILD.bazel
@@ -208,6 +208,7 @@ py_test_module_list(
         "test_scheduling.py",
         "test_scheduling_2.py",
         "test_token_auth_integration.py",
+        "test_token_auth_integration_2.py",
         "test_wait.py",
     ],
     name_suffix = "_client_mode",
@@ -351,6 +352,7 @@ py_test_module_list(
         "test_tempdir.py",
         "test_tls_auth.py",
         "test_token_auth_integration.py",
+        "test_token_auth_integration_2.py",
         "test_traceback.py",
         "test_worker_capping.py",
         "test_worker_state.py",
